+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## NEXT
+## 0.2.5
 
 - Added the `cai_causal_graph.identify_utils.identify_confounders` utility function, which allows you to identify
   a list of confounders between two nodes in a `cai_causal_graph.causal_graph.CausalGraph`.